Added a check to not cause error

+++ Main.py
@@ -13,8 +13,8 @@ gameName, users, Placed, Location = None, None, None, None
 def getLocation():
     return input("Enter location to place ship: ")
 
-
-os.remove('Tests/Path.txt')  # removes old data at start
+if os.path.exists('Tests/Path.txt'):
+	os.remove('Tests/Path.txt')  # removes old data at start
 
 
 while True:
